requirements.txt
moviepy==1.0.3
openai-whisper==20230314
fastapi==0.95.0
uvicorn==0.22.0
faiss-cpu==1.7.3
sentence-transformers==2.2.2
torch==2.0.1
docker==6.0.1
kubernetes==26.1.0
dvc==3.0.0
mlflow==2.4.0
prometheus-client==0.16.0
grafana-api==1.0.0
yt-dlp==2023.3.4